requirements.txt
streamlit
langchain-groq
python-dotenv
pymupdf
pandas